+++ bin/nanocld/src/utils/vm.rs
@@ -211,7 +211,7 @@ pub async fn create_instance(
   }
 
   let config = bollard_next::container::Config {
-    image: Some("nanocl-qemu:0.1.0".into()),
+    image: Some("ghcr.io/nxthat/nanocl-qemu:7.1.0.0".into()),
     tty: Some(true),
     hostname: vm.config.hostname.clone(),
     env: Some(env),