+++ bin/nanocld/changelog.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [untagged]
+
+### Fixed
+
+- Missing `created_at` field for resources
+- Order by `created_at` by default for resources
+
 ## [0.6.1] - 2023-05-14
 
 ### Fixed