+++ bin/nanocld/changelog.md
@@ -5,11 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [untagged]
+## [0.7.0] - 2023-05-22
 
 ### Changed
 
 - Vm runtime image to `ghcr.io/nxthat/nanocl-qemu:7.1.0.0`
+- Better state messages
 
 ### Added